+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Fixed a bug where go refiner would fail with a null reference.
+
 ## [0.11.1] - 2023-02-13
 
 ### Added